requirements.txt
streamlit==1.45.1
langchain==0.3.25
langchain-community==0.3.24
langchain-huggingface==0.2.0
faiss-cpu==1.11.0
python-docx==1.1.2
pypdf==5.5.0
pandas==2.2.3
tiktoken==0.9.0
requests==2.32.3
python-dotenv==1.1.0
huggingface-hub==0.32.3
xlrd==2.0.1
openpyxl==3.1.5
openai==1.82.1
sentence-transformers==4.1.0
transformers==4.52.4
scikit-learn==1.6.1
torch==2.7.0
PyMuPDF-1.26.0